Add WASM binding
This adds a WASM binding written in C++ that exposes a `simulate()`
function to WASM runtimes. The world state is exchanged via FlatBuffers
and all required memory allocation happens inside the WASM module.

From now on, the WASM module is built for new releases and added to the
artifacts.

Further, a new function is added to the C-API:
`twsfwphysx_clear_missile_batch()` that clears the missile batch without
deallocating the memory. (This function was needed to implement the WASM
module.)